// api/src/database/typings/Models.d.ts
export interface Models {
  Tenant: typeof Tenant;
  User: typeof User;
  UserProfile: typeof UserProfile;
  Role: typeof Role;
  Permission: typeof Permission;
  Client: typeof Client;
  ServiceAddress: typeof ServiceAddress;
  ServiceItemTemplate: typeof ServiceItemTemplate;
  Service: typeof Service;
  ServiceItem: typeof ServiceItem;
  ServiceSkill: typeof ServiceSkill;
  Schedule: typeof Schedule;
  Job: typeof Job;
  JobNote: typeof JobNote;
  JobDocument: typeof JobDocument;
  JobExpenses: typeof JobExpenses;
  JobExpensesItem: typeof JobExpensesItem;
  JobHistory: typeof JobHistory;
  Vehicle: typeof Vehicle;
  Entity: typeof Entity;
  Invoice: typeof Invoice;
  ServiceTemplate: typeof ServiceTemplate;
  SkillTemplate: typeof SkillTemplate;
  UserSkill: typeof UserSkill;
  ChecklistTemplate: typeof ChecklistTemplate;
  ChecklistItemTemplate: typeof ChecklistItemTemplate;
  ChecklistJob: typeof ChecklistJob;
  ChecklistJobItem: typeof ChecklistJobItem;
  Setting: typeof Setting;
  TableColumnSetting: typeof TableColumnSetting;
  AppLog: typeof AppLog;
  Agent: typeof Agent;
  ContactPerson: typeof ContactPerson;
  Rating: typeof Rating;
  JobNoteTemplate: typeof JobNoteTemplate;
  Equipment: typeof Equipment;
  JobLabelTemplate: typeof JobLabelTemplate;
  JobLabel: typeof JobLabel;
  WaJob: typeof WaJob;
  Notification: typeof Notification;
  BrandTemplate: typeof BrandTemplate;
  ClientDocument: typeof ClientDocument;
  GstTemplate: typeof GstTemplate;
  InvoiceHistory: typeof InvoiceHistory;
  CollectedAmountHistory: typeof CollectedAmountHistory;
  JobNoteMedia: typeof JobNoteMedia;
  CustomField: typeof CustomField;
  TimeOff: typeof TimeOff;
  District: typeof District;
  RoleGrant: typeof RoleGrant;
  PdfTemplateOptions: typeof PdfTemplateOptions;
  BookingSetting: typeof BookingSetting;
  [key: string]: typeof ModelBase;
}